[Text Directionality] Move the text directionality code to separate files https://bugs.webkit.org/show_bug.cgi?id=278698 rdar://134754984
Reviewed by Ryosuke Niwa.

This is a follow-up of 282648@main. This will make the element text direction
code easier to read and debug.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/dom/Element.cpp:
(WebCore::parseTextDirectionState): Deleted.
(WebCore::elementTextDirectionState): Deleted.
(WebCore::elementHasValidTextDirectionState): Deleted.
(WebCore::elementHasAutoTextDirectionState): Deleted.
(WebCore::updateHasDirAutoFlagForSubtree): Deleted.
(WebCore::updateElementHasDirAutoFlag): Deleted.
(WebCore::computeTextDirectionFromText): Deleted.
(WebCore::computeTextDirection): Deleted.
(WebCore::computeContainedTextAutoDirection): Deleted.
(WebCore::computeTextDirectionOfSlotElement): Deleted.
(WebCore::updateEffectiveTextDirectionOfElementAndShadowTree): Deleted.
(WebCore::updateEffectiveTextDirectionState): Deleted.
(WebCore::Element::computeAutoTextDirection const): Deleted.
(WebCore::Element::computeTextDirectionIfDirIsAuto const): Deleted.
(WebCore::Element::updateEffectiveTextDirectionOfDescendants): Deleted.
(WebCore::Element::updateEffectiveTextDirectionOfAncestors): Deleted.
* Source/WebCore/dom/Element.h:
* Source/WebCore/dom/ElementTextDirection.cpp: Added.
(WebCore::parseTextDirectionState):
(WebCore::elementTextDirectionState):
(WebCore::elementHasValidTextDirectionState):
(WebCore::elementHasAutoTextDirectionState):
(WebCore::updateHasDirAutoFlagForSubtree):
(WebCore::updateElementHasDirAutoFlag):
(WebCore::computeTextDirectionFromText):
(WebCore::computeTextDirection):
(WebCore::computeContainedTextAutoDirection):
(WebCore::computeTextDirectionOfSlotElement):
(WebCore::computeAutoDirectionality):
(WebCore::computeTextDirectionIfDirIsAuto):
(WebCore::updateEffectiveTextDirectionOfElementAndShadowTree):
(WebCore::updateEffectiveTextDirectionState):
(WebCore::updateEffectiveTextDirectionOfDescendants):
(WebCore::updateEffectiveTextDirectionOfAncestors):
* Source/WebCore/dom/ElementTextDirection.h: Added.
* Source/WebCore/html/HTMLTextFormControlElement.cpp:
(WebCore::HTMLTextFormControlElement::directionForFormData const):
* Source/WebCore/html/InputType.cpp:
(WebCore::InputType::setValue):
* Source/WebCore/style/ElementRuleCollector.cpp:
(WebCore::Style::ElementRuleCollector::matchAllRules):

Canonical link: https://commits.webkit.org/282796@main
