+++ Source/WebCore/Sources.txt
@@ -1171,6 +1171,7 @@ dom/Element.cpp
 dom/ElementData.cpp
 dom/ElementInternals.cpp
 dom/ElementRareData.cpp
+dom/ElementTextDirection.cpp
 dom/ErrorEvent.cpp
 dom/Event.cpp
 dom/EventContext.cpp
